refactor: PieceFactory 추출
