+++ src/main/java/chess/domains/piece/Piece.java
@@ -3,9 +3,6 @@ package chess.domains.piece;
 import chess.domains.position.Direction;
 import chess.domains.position.Position;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Objects;
 
 public abstract class Piece {
@@ -21,60 +18,16 @@ public abstract class Piece {
     private static final String INVALID_DIAGONAL_MOVE_OF_PAWN_ERR_MSG = "폰은 상대말을 잡는 경우 이 외에 대각선으로 이동할 수 없습니다.";
     private static final String INVALID_VERTICAL_MOVE_OF_PAWN_ERR_MSG = "폰은 앞에 있는 상대를 잡을 수 없습니다.";
     private static final String INVALID_MOVE_OPPONENT_PIECE_ERR_MSG = "상대방의 말을 움직일 수 없습니다.";
-    private static final List<Piece> whitePieces;
-    private static final List<Piece> whitePawnsPieces;
-    private static final List<Piece> blackPieces;
-    private static final List<Piece> blackPawnsPieces;
-    private static final List<Piece> blankPieces;
 
     protected final PieceColor pieceColor;
     private final PieceType type;
-
-    static {
-        whitePieces = createBundleByColor(PieceColor.WHITE);
-        whitePawnsPieces = createPawnsBundleByColor(PieceColor.WHITE);
-        blackPieces = createBundleByColor(PieceColor.BLACK);
-        blackPawnsPieces = createPawnsBundleByColor(PieceColor.BLACK);
-        blankPieces = createBlankBundle();
-    }
 
     public Piece(PieceColor pieceColor, PieceType type) {
         this.pieceColor = pieceColor;
         this.type = type;
     }
 
-    private static List<Piece> createBundleByColor(PieceColor color) {
-        return new ArrayList<>(Arrays.asList(
-                new Rook(color), new Knight(color), new Bishop(color), new King(color),
-                new Queen(color), new Bishop(color), new Knight(color), new Rook(color)));
-    }
-
-
-    private static List<Piece> createPawnsBundleByColor(PieceColor color) {
-        return new ArrayList<>(Arrays.asList(
-                new Pawn(color), new Pawn(color), new Pawn(color), new Pawn(color),
-                new Pawn(color), new Pawn(color), new Pawn(color), new Pawn(color)));
-    }
-
-    private static List<Piece> createBlankBundle() {
-        return new ArrayList<>(Arrays.asList(
-                new Blank(), new Blank(), new Blank(), new Blank(),
-                new Blank(), new Blank(), new Blank(), new Blank()));
-    }
-
     public abstract boolean isValidMove(Position currentPosition, Position targetPosition);
-
-    public static List<Piece> getBlackPieces() {
-        return blackPieces;
-    }
-
-    public static List<Piece> getWhitePieces() {
-        return whitePieces;
-    }
-
-    public static List<Piece> getWhitePawnsPieces() {
-        return whitePawnsPieces;
-    }
 
     public boolean isMine(PieceColor pieceColor) {
         return this.pieceColor == pieceColor;
@@ -97,14 +50,6 @@ public abstract class Piece {
 
     public double score() {
         return type.getScore();
-    }
-
-    public static List<Piece> getBlackPawnsPieces() {
-        return blackPawnsPieces;
-    }
-
-    public static List<Piece> getBlankPieces() {
-        return blankPieces;
     }
 
     public void checkSameColorWith(PieceColor teamColor) {
